+++ bindings/nostr-sdk-ffi/src/database/mod.rs
@@ -17,7 +17,7 @@ pub mod events;
 use self::events::Events;
 use crate::error::Result;
 use crate::protocol::event::Event;
-use crate::protocol::types::Filter;
+use crate::protocol::filter::Filter;
 
 /// Reason why event wasn't stored into the database
 #[derive(Enum)]