+++ bindings/nostr-sdk-ffi/src/protocol/message/client.rs
@@ -10,7 +10,7 @@ use uniffi::{Enum, Object};
 
 use crate::error::Result;
 use crate::protocol::event::Event;
-use crate::protocol::types::filter::Filter;
+use crate::protocol::filter::Filter;
 
 /// Messages sent by clients, received by relays
 #[derive(Enum)]