+++ crates/nostr/src/prelude.rs
@@ -24,6 +24,7 @@ pub use crate::event::kind::{self, *};
 pub use crate::event::tag::{self, *};
 pub use crate::event::unsigned::{self, *};
 pub use crate::event::{self, *};
+pub use crate::filter::{self, *};
 pub use crate::key::{self, *};
 pub use crate::message::{self, *};
 // NIPs