+++ crates/nostr/src/types/mod.rs
@@ -8,14 +8,12 @@
 #![allow(ambiguous_glob_reexports)]
 
 pub mod contact;
-pub mod filter;
 pub mod image;
 pub mod metadata;
 pub mod time;
 pub mod url;
 
 pub use self::contact::*;
-pub use self::filter::*;
 pub use self::image::*;
 pub use self::metadata::*;
 pub use self::time::*;